delete MinerFromPlugin merge into Miner class
