+++ src/NiceHashMiner/Mining/Miner.cs
@@ -10,7 +10,7 @@ using NiceHashMiner.Mining.Plugins;
 
 namespace NiceHashMiner.Mining
 {
-    public abstract class Miner
+    public class Miner
     {
         public static Miner CreateMinerForMining(List<MiningPair> miningPairs, string groupKey)
         {
@@ -20,7 +20,7 @@ namespace NiceHashMiner.Mining
             var plugin = MinerPluginsManager.GetPluginWithUuid(algorithm.MinerID);
             if (plugin != null)
             {
-                return new MinerFromPlugin(plugin, miningPairs, groupKey);
+                return new Miner(plugin, miningPairs, groupKey);
             }
             return null;
         }
@@ -45,16 +45,25 @@ namespace NiceHashMiner.Mining
 
         public bool IsUpdatingApi { get; protected set; } = false;
 
+        // Now every single miner is based from the Plugins
+        private readonly PluginContainer _plugin;
+        private readonly IMiner _miner;
 
-        protected Miner(string minerDeviceName, List<MiningPair> miningPairs, string groupKey)
+        private readonly SemaphoreSlim _apiSemaphore = new SemaphoreSlim(1, 1);
+
+        // you must use 
+        protected Miner(PluginContainer plugin, List<MiningPair> miningPairs, string groupKey)
         {
+            _plugin = plugin;
+            _miner = _plugin.CreateMiner();
+
             MiningPairs = miningPairs;
             IsInit = MiningPairs != null && MiningPairs.Count > 0;
             if (IsInit)
             {
                 foreach (var pair in miningPairs)
                 {
-                    // for PRODUCTION we still need these indexes get rid of this when possible
+                    // for OLD PRODUCTION we still need these indexes get rid of this when possible
                     var cDev = AvailableDevices.GetDeviceWithUuidOrB64Uuid(pair.Device.UUID);
                     var index = cDev?.Index ?? -1;
                     if (index < 0) continue;
@@ -64,7 +73,7 @@ namespace NiceHashMiner.Mining
 
             GroupKey = groupKey;
 
-            MinerDeviceName = minerDeviceName;
+            MinerDeviceName = plugin.PluginUUID;
             Logger.Info(MinerTag(), "NEW MINER CREATED");
         }
 
@@ -88,10 +97,101 @@ namespace NiceHashMiner.Mining
             return _minerTag;
         }
 
-        public abstract void Start(string miningLocation, string username);
-        public abstract Task<ApiData> GetSummaryAsync();
-        public abstract void Stop();
+        public async Task<ApiData> GetSummaryAsync()
+        {
+            var apiData = new ApiData();
+            if (!IsUpdatingApi)
+            {
+                IsUpdatingApi = true;
+                await _apiSemaphore.WaitAsync();
+                try
+                {
+                    apiData = await _miner.GetMinerStatsDataAsync();
+                }
+                catch (Exception e)
+                {
+                    Logger.Error(MinerTag(), $"GetSummaryAsync error: {e.Message}");
+                }
+                finally
+                {
+                    IsUpdatingApi = false;
+                    _apiSemaphore.Release();
+                }
+            }
 
+            UpdateApiTimestamp(apiData);
+
+            // TODO workaround plugins should return this info
+            // create empty stub if it is null
+            if (apiData == null)
+            {
+                Logger.Debug(MinerTag(), "GetSummary returned null... Will create ZERO fallback");
+                apiData = new ApiData();
+            }
+            if (apiData.AlgorithmSpeedsPerDevice == null)
+            {
+                apiData = new ApiData();
+                var perDevicePowerDict = new Dictionary<string, int>();
+                var perDeviceSpeedsDict = new Dictionary<string, IReadOnlyList<AlgorithmTypeSpeedPair>>();
+                var perDeviceSpeeds = MiningPairs.Select(pair => (pair.Device.UUID, pair.Algorithm.IDs.Select(type => new AlgorithmTypeSpeedPair(type, 0d))));
+                foreach (var kvp in perDeviceSpeeds)
+                {
+                    var uuid = kvp.Item1; // kvp.UUID compiler doesn't recognize ValueTypes lib???
+                    perDeviceSpeedsDict[uuid] = kvp.Item2.ToList();
+                    perDevicePowerDict[uuid] = 0;
+                }
+                apiData.AlgorithmSpeedsPerDevice = perDeviceSpeedsDict;
+                apiData.PowerUsagePerDevice = perDevicePowerDict;
+                apiData.PowerUsageTotal = 0;
+                apiData.AlgorithmSpeedsTotal = perDeviceSpeedsDict.First().Value;
+            }
+            else if (apiData.AlgorithmSpeedsPerDevice != null && apiData.PowerUsagePerDevice.Count == 0)
+            {
+                var perDevicePowerDict = new Dictionary<string, int>();
+                foreach (var kvp in MiningPairs)
+                {
+                    var uuid = kvp.Device.UUID;
+                    perDevicePowerDict[uuid] = 0;
+                }
+                apiData.PowerUsagePerDevice = perDevicePowerDict;
+                apiData.PowerUsageTotal = 0;
+            }
+
+            // TODO temporary here move it outside later
+            MiningStats.UpdateGroup(apiData, _plugin.PluginUUID, _plugin.Name);
+
+            return apiData;
+        }
+
+        // TODO this thing 
+        public void Start(string miningLocation, string username)
+        {
+            if (_isEnded) return;
+            _miner.InitMiningLocationAndUsername(miningLocation, username);
+            _miner.InitMiningPairs(MiningPairs);
+            EthlargementIntegratedPlugin.Instance.Start(MiningPairs);
+            _miner.StartMining();
+            IsRunning = true;
+            // maxTimeout = ConfigManager.GeneralConfig.CoolDownCheckEnabled
+            var maxTimeout = _plugin.GetApiMaxTimeout(MiningPairs);
+            MinerApiWatchdog.AddGroup(GroupKey, maxTimeout, DateTime.UtcNow);
+        }
+
+        public void Stop()
+        {
+            // TODO thing about this case, closing opening on switching
+            // EthlargementIntegratedPlugin.Instance.Stop(_miningPairs);
+            MinerApiWatchdog.RemoveGroup(GroupKey);
+            MiningStats.RemoveGroup(MiningPairs.Select(pair => pair.Device.UUID), _plugin.PluginUUID);
+            IsRunning = false;
+            _miner.StopMining();
+            //if (_miner is IDisposable disposableMiner)
+            //{
+            //    disposableMiner.Dispose();
+            //}
+        }
+
+        // TODO check cleanup here
         public void End()
         {
             try
@@ -157,5 +257,43 @@ namespace NiceHashMiner.Mining
                 }
             }
         }
+
+
+        #region MinerApiWatchdog
+        private double _lastSpeedsTotalSum = 0d;
+        private double _lastPerDevSpeedsTotalSum = 0d;
+
+        // TODO this can be moved in MinerApiWatchdog
+        private void UpdateApiTimestamp(ApiData apiData)
+        {
+            // we will not update api timestamps if we have no data or speeds are zero
+            if (apiData == null)
+            {
+                // TODO debug log no api data
+                return;
+            }
+            if (apiData.AlgorithmSpeedsTotal == null && apiData.AlgorithmSpeedsPerDevice == null)
+            {
+                // TODO debug log cannot get speeds
+                return;
+            }
+            var speedsTotalSum = apiData.AlgorithmSpeedsTotal?.Select(p => p.Speed).Sum() ?? 0d;
+            var perDevSpeedsTotalSum = apiData.AlgorithmSpeedsPerDevice?.Values.SelectMany(pl => pl).Select(p => p.Speed).Sum() ?? 0d;
+            if (speedsTotalSum == 0d && perDevSpeedsTotalSum == 0d)
+            {
+                // TODO debug log speeds are zero
+                return;
+            }
+            if (speedsTotalSum == _lastSpeedsTotalSum && perDevSpeedsTotalSum == _lastPerDevSpeedsTotalSum)
+            {
+                // TODO debug log speeds seem to be stuck
+                return;
+            }
+            // update 
+            _lastSpeedsTotalSum = speedsTotalSum;
+            _lastPerDevSpeedsTotalSum = perDevSpeedsTotalSum;
+            MinerApiWatchdog.UpdateApiTimestamp(GroupKey, DateTime.UtcNow);
+        }
+        #endregion MinerApiWatchdog
     }
 }
